refactor: 미사용 dto 제거, dto 명명 수정
