--- src/main/java/com/wanted/preonboarding/ticket/domain/dto/MessagesDto.java
@@ -1,14 +1,0 @@
-package com.wanted.preonboarding.ticket.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter @Setter
-public class MessagesDto {
-    private String to;
-    private String content;
-}
